Update README.md
There was a mistake in the documentation. The SendSMS.send() function accepts a callback function as its third parameter. This callback function receives two parameters: the first is the message ID (which is the same variable as the first parameter of the SendSMS.send() function), and the second is the status of the sent SMS.